// src/app/data-models/employee-type.ts
export type Employee = {
    _id: string;
    first_name: string;
    last_name: string;
    email: string;
    gender: string;
    designation: string;
    salary: number;
    date_of_joining: string;
    department: string;
    employee_photo: string;
    created_at: string;
    updated_at: string;
  };